CI: Fix workflow input for notarization
I thought the type was just for the UI and the values were always strings (maybe it used to be that way?), but apparently not. They are legit bools, well more like nullable bools since they can be undefined if the run is triggered automatically (e.g. via a push) regardless of whether a default is specified.