Fix regression with macOS session restore
The logical pixel rounding code made this function return a different value when a pixmap isn't loaded, confusing one of the validation checks in ScrollHelper. This made it sometimes attempt to constrain the position even if an image wasn't loaded, causing the scroll position not to restore properly in some cases if the titlebar is hidden.